requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
numpy==1.26.3
pandas==2.1.4
psycopg==3.1.15
psycopg2==2.9.9
python-dateutil==2.8.2
python-poppler==0.4.1
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
